requirements.txt
Babel==2.1.1
Chameleon==2.22
Mako==1.0.3
MarkupSafe==0.23
PasteDeploy==1.5.2
Pygments==2.0.2
SQLAlchemy==1.0.8
WebOb==1.5.0a1
WebTest==2.0.20
alembic==0.8.3
argparse==1.1
bcrypt==2.0.0
beautifulsoup4==4.4.1
cffi==1.3.1
debtcollector==1.0.0
dogpile.cache==0.6.1
futures==3.0.3
gunicorn==19.4.2
iso8601==0.1.11
monotonic==0.4
msgpack-python==0.4.6
netaddr==0.7.18
netifaces==0.10.4
nltk==3.0.5
nose==1.3.7
numpy==1.10.1
oslo.config==3.0.0
oslo.i18n==3.0.0
oslo.serialization==2.0.0
oslo.utils==3.0.0
passlib==1.6.5
pbr==1.8.1
prettytable==0.7.2
psycopg2==2.6.1
pycparser==2.14
pympi-ling==1.40
pyramid==1.6a2
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.4.1
pyramid-mako==1.0.2
pyramid-tm==0.12
python-editor==0.4
python-keystoneclient==1.8.1
python-memcached==1.58
python-swiftclient==2.6.0
pytz==2015.7
repoze.lru==0.6
requests==2.7.0
simplejson==3.8.0
six==1.10.0
stevedore==1.9.0
tgt==1.3.1
transaction==1.4.4
translationstring==1.3
venusian==1.0
waitress==0.8.9
wrapt==1.10.5
werkzeug==0.11.4
zope.deprecation==4.1.2
zope.interface==4.1.2
zope.sqlalchemy==0.7.6
